quick-start/requirements.txt
openai
langchain
python-dotenv
google-search-results